requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
PyYAML